+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This project follows [Semantic Versioning](https://semver.org/).
 - **Batched mode** - simmilar to async and chained, **batch** mode passes an input by batch instead of by step. The resulting batch output is merged and passed into the next batch.
 - **Concurrency** - as a precaution, _all workflows_ go through the same concurrency handler that **async** uses, **chained** and **fanout** modes are forced to use a concurrency of 1.
 
+### Fixed
+
+- **Concurrency override** - concurrency was being overridden by default value in flags when using `$ xgsd run`. This has been fixed.
+
 ## [`v0.4.1`] - 2025-09-08
 
 ### Fixed